refactor(#2824): move the prefetch-advice pins to their own module (campsite rule)
The file-size ratchet FAILs when a change grows an already-over-threshold file.
reader/tests.rs (1135/800) would have grown by 48 lines, so the retargeted #1143
pin and the new default-config pin move to reader/prefetch_advice_tests.rs, a new
90-line module gated #[cfg(all(test, unix))] at its declaration — the same
in-crate pattern read_at_point_tests.rs uses, and for the same reason
(mmap_advice_for is pub(super)). reader/tests.rs now SHRINKS by 40 lines.

The three other over-threshold files (reader/mod.rs, data_access/mod.rs,
scan_stream_windowed.rs) had grown 1-2 lines purely from comment rewrap; those
comments are reflowed back to their base line counts (mod.rs is -1).

Refs #2824, #1116, #1135
