+++ cqlite-core/src/storage/sstable/reader/mod.rs
@@ -40,6 +40,8 @@ pub(crate) mod partition_successor;
 mod fd_gauge;
 mod read_at;
 // Concurrency scenarios for the ReadAt point-read migration (issue #1573).
+#[cfg(all(test, unix))] // mmap advice: #2824 WILLNEED + the #1143 Sequential ban
+mod prefetch_advice_tests;
 #[cfg(test)]
 mod read_at_point_tests;
 // Direct-I/O read-ahead window sizing (issue #1596, Epic F / F6): clamps a
@@ -159,25 +161,20 @@ pub(crate) fn index_db_sibling(data_path: &Path) -> Option<PathBuf> {
     Some(parent.join(format!("{base}-Index.db")))
 }
 
-/// Minimum SSTable size for the point-read path to get its OWN mmap advised
-/// `MADV_RANDOM` (issue #2210). Below this the point source shares the scan's
-/// `Arc<Mmap>` unchanged (no 2nd mapping, so it carries whatever
-/// `mmap_advice_for` gave the scan plane): the whole file is small
-/// enough that the kernel's default read-ahead cheaply makes it resident, and
-/// `MADV_RANDOM` would only add per-page synchronous faults. Above it, scattered
-/// point-lookup faults otherwise waste the ~128 KiB read-ahead window per read,
-/// so a dedicated `MADV_RANDOM` mapping collapses both the block-I/O
-/// amplification (~30x) and the cold-cache per-read latency (~35-43%). Threshold
-/// is measurement-derived on Linux/EBS: the win is unambiguous by 4 MiB; 8 MiB
-/// leaves 2x margin above the sub-MB "wash" zone. See
-/// docs/reports/issue-2210-madv-random-point-mmap-ab.md. The SCAN mapping never
-/// gets `MADV_RANDOM`; under the default `PrefetchMode::Auto` it gets
-/// `MADV_WILLNEED` (asynchronous read-ahead, no drop-behind — issue #2824),
-/// `Off` leaves it unadvised, and an explicit `Sequential`/`WillNeed` is honoured
-/// as named. It backs BOTH `BlockSource::Mapped` and `scan_positional_source`,
-/// and the windowed / Summary-guided scan feed reads through the latter
-/// (`ReadAt`), NOT through `BlockSource` (#2876) — so a claim scoped to
-/// `BlockSource` describes nothing that a real scan does.
+/// Minimum SSTable size for the point-read path to get its OWN mmap advised `MADV_RANDOM` (issue
+/// #2210). Below this the point source shares the scan's `Arc<Mmap>` unchanged (no 2nd mapping; it
+/// carries the scan plane's advice): the whole file is small enough that read-ahead cheaply makes
+/// it resident and `MADV_RANDOM` would only add per-page faults. Above it, scattered point-lookup
+/// faults otherwise waste the ~128 KiB read-ahead window per read, so a dedicated `MADV_RANDOM`
+/// mapping collapses both the block-I/O amplification (~30x) and the cold-cache per-read latency
+/// (~35-43%). Threshold is measurement-derived on Linux/EBS: the win is unambiguous by 4 MiB; 8
+/// MiB leaves 2x margin above the sub-MB "wash" zone. See
+/// docs/reports/issue-2210-madv-random-point-mmap-ab.md. The SCAN mapping never gets
+/// `MADV_RANDOM`; the default `PrefetchMode::Auto` advises it `MADV_WILLNEED` (async read-ahead,
+/// no drop-behind — #2824), `Off` leaves it unadvised, and `Sequential`/`WillNeed` are honoured as
+/// named. It backs BOTH `BlockSource::Mapped` and `scan_positional_source`, and the windowed /
+/// Summary-guided scan feed reads through the latter (`ReadAt`), NOT through `BlockSource` (#2876)
+/// — so a claim scoped to `BlockSource` describes nothing that a real scan does.
 #[cfg(unix)]
 const POINT_MMAP_MADV_RANDOM_MIN_BYTES: u64 = 8 * 1024 * 1024;
